231201 회원탈퇴시 권한 확인

+++ src/main/java/com/example/prj2be231123/controller/MemberController.java
@@ -77,7 +77,16 @@ public class MemberController {
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
         } else {
